+++ source/main/resources/tasks/MSBuild.XCode/MSBuild.XCode/PackageRepository/Remote/Databases/IPackageDatabase.cs
@@ -7,11 +7,11 @@ namespace xpackage_repo
 {
     public interface IPackageDatabase
     {
-        bool submit(PackageVersion_pv package, List<KeyValuePair<string, int>> dependencies);
+        bool submit(PackageVersion_pv package, List<KeyValuePair<string, Int64>> dependencies);
 
         bool findUniqueVersion(PackageVersion_pv package);
-        bool findLatestVersion(PackageVersion_pv package, out int outVersion);
-        bool findLatestVersion(PackageVersion_pv package, int start_version, bool include_start, int end_version, bool include_end, out int outVersion);
+        bool findLatestVersion(PackageVersion_pv package, out Int64 outVersion);
+        bool findLatestVersion(PackageVersion_pv package, Int64 start_version, bool include_start, Int64 end_version, bool include_end, out Int64 outVersion);
 
         bool retrieveVarsOf(PackageVersion_pv package, out Dictionary<string, object> vars);
     }
